requirements.txt
Django==2.1.8
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-imagekit==4.0.2
pilkit==2.0
Pillow==6.0.0
pytz==2019.1
six==1.12.0
